feat: rename DXT to MCPB and update related upload functionality
